reorganize blocks
