--- rpc/backend/blocks.go
@@ -1,99 +1,0 @@
-package backend
-
-import (
-	"fmt"
-	"strconv"
-
-	grpctypes "github.com/cosmos/cosmos-sdk/types/grpc"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/evmos/ethermint/rpc/types"
-	evmtypes "github.com/evmos/ethermint/x/evm/types"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
-)
-
-// Getting Blocks
-//
-// Retrieves information from a particular block in the blockchain.
-// BlockNumber() (hexutil.Uint64, error)
-// GetBlockByNumber(ethBlockNum rpctypes.BlockNumber, fullTx bool) (map[string]interface{}, error)
-// GetBlockByHash(hash common.Hash, fullTx bool) (map[string]interface{}, error)
-
-// BlockNumber returns the current block number in abci app state.
-// Because abci app state could lag behind from tendermint latest block, it's more stable
-// for the client to use the latest block number in abci app state than tendermint rpc.
-func (b *Backend) BlockNumber() (hexutil.Uint64, error) {
-	// do any grpc query, ignore the response and use the returned block height
-	var header metadata.MD
-	_, err := b.queryClient.Params(b.ctx, &evmtypes.QueryParamsRequest{}, grpc.Header(&header))
-	if err != nil {
-		return hexutil.Uint64(0), err
-	}
-
-	blockHeightHeader := header.Get(grpctypes.GRPCBlockHeightHeader)
-	if headerLen := len(blockHeightHeader); headerLen != 1 {
-		return 0, fmt.Errorf("unexpected '%s' gRPC header length; got %d, expected: %d", grpctypes.GRPCBlockHeightHeader, headerLen, 1)
-	}
-
-	height, err := strconv.ParseUint(blockHeightHeader[0], 10, 64)
-	if err != nil {
-		return 0, fmt.Errorf("failed to parse block height: %w", err)
-	}
-
-	return hexutil.Uint64(height), nil
-}
-
-// GetBlockByNumber returns the block identified by number.
-func (b *Backend) GetBlockByNumber(blockNum types.BlockNumber, fullTx bool) (map[string]interface{}, error) {
-	resBlock, err := b.GetTendermintBlockByNumber(blockNum)
-	if err != nil {
-		return nil, nil
-	}
-
-	// return if requested block height is greater than the current one
-	if resBlock == nil || resBlock.Block == nil {
-		return nil, nil
-	}
-
-	blockRes, err := b.GetTendermintBlockResultByNumber(&resBlock.Block.Height)
-	if err != nil {
-		b.logger.Debug("failed to fetch block result from Tendermint", "height", blockNum, "error", err.Error())
-		return nil, nil
-	}
-
-	res, err := b.EthBlockFromTendermint(resBlock, blockRes, fullTx)
-	if err != nil {
-		b.logger.Debug("EthBlockFromTendermint failed", "height", blockNum, "error", err.Error())
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// GetBlockByHash returns the block identified by hash.
-func (b *Backend) GetBlockByHash(hash common.Hash, fullTx bool) (map[string]interface{}, error) {
-	resBlock, err := b.GetTendermintBlockByHash(hash)
-	if err != nil {
-		return nil, err
-	}
-
-	if resBlock == nil {
-		// block not found
-		return nil, nil
-	}
-
-	blockRes, err := b.GetTendermintBlockResultByNumber(&resBlock.Block.Height)
-	if err != nil {
-		b.logger.Debug("failed to fetch block result from Tendermint", "block-hash", hash.String(), "error", err.Error())
-		return nil, nil
-	}
-
-	res, err := b.EthBlockFromTendermint(resBlock, blockRes, fullTx)
-	if err != nil {
-		b.logger.Debug("EthBlockFromTendermint failed", "hash", hash, "error", err.Error())
-		return nil, err
-	}
-
-	return res, nil
-}
